pyproject.toml
[project]
name = "image-dedup"
version = "2.0.0"
description = "Desktop GUI image deduplication tool with forensic analysis"
requires-python = ">=3.12"
dependencies = [
    "PyQt6>=6.6",
    "opencv-python-headless>=4.9",
    "Pillow>=10.0",
    "imagehash>=4.3",
    "numpy>=1.26",
    "scipy>=1.12",
    "Jinja2>=3.1",
    "openpyxl>=3.1",
    "PyMuPDF>=1.24",
    "rarfile>=4.1",
    "py7zr>=0.21",
    "pillow-heif>=0.16",
    "pillow-avif-plugin>=1.4",
    "reportlab>=4.1",
]

[project.optional-dependencies]
ai = ["transformers>=4.40", "torch>=2.2"]

[project.scripts]
image-dedup = "image_dedup.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_dedup"]
